Remove ginvite from our modules/custom folder
We have used a patch to add ginvite to our distribution ever since 10.x, trying to update that to D9 compatibility
lead us to fork the patch and add it as custom module. Since then we've noticed the community started to use a contrib module,
we want to move away from this custom fork and instead use the contrib module.